requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
Django==5.2.9
django-cloudinary-storage==0.3.0
dj-database-url==2.2.0
django-cors-headers==4.9.0
gunicorn==22.0.0
idna==3.11
MarkupSafe==3.0.3
pillow==11.3.0
psycopg[binary]==3.2.6
pycparser==2.23
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.3
whitenoise==6.7.0
Werkzeug==3.1.5
